version2 for testing requests
Now, in the main function, you can set as many strings as you want, and
transfer them into the member TestNetworkStream variable in WebServer
class. This version works well with many tests of Web request you want.
In fact, the stream.write is not implemented. So, if you get not
implemented error, then it means you are sending a valid http request.
Otherwise, the program will just quit.